pyproject.toml
[project]
name = "odoo-data-flow"
version = "0.0.1"
description = "Odoo Data Flow"
readme = "README.md"
requires-python = ">=3.9"
license = "LGPL-3.0-only"
authors = [
    { name = "bosd", email = "c5e2fd43-d292-4c90-9d1f-74ff3436329a@anonaddy.me" },
]
classifiers = ["Development Status :: 3 - Alpha"]

dependencies = [
    "click >=8.0.1",
    "odoo-client-lib @ git+https://github.com/odoo/odoo-client-lib.git@refs/pull/5/head",
    "httpx",
    "lxml",
    "rich",
    "polars",
    "pyarrow",
]


[project.urls]
Homepage = "https://github.com/OdooDataFlow/odoo-data-flow"
Repository = "https://github.com/OdooDataFlow/odoo-data-flow"
Documentation = "https://odoo-data-flow.readthedocs.io"
Changelog = "https://github.com/OdooDataFlow/odoo-data-flow/releases"

# --- Added for mypyc integration ---
[build-system]
# This now includes all dependencies needed for mypyc to compile the code.
requires = [
    "setuptools>=61.0",
    "mypy>=1.0",
    "click >=8.0.1",
    "odoo-client-lib",
    "httpx",
    "lxml",
    "rich",
    "polars",
    "pyarrow",
]
build-backend = "setuptools.build_meta"

# This section is the standard for defining extras for pip/uv install
[project.optional-dependencies]
dev = [
    "coverage[toml] >= 6.2",
    "pre-commit >=2.16.0",
    "pre-commit-hooks >=4.6.0",
    "pytest >=6.2.5",
    "pygments >=2.10.0",
    "nox >=2024.04.14",
    "pytest-mock",
]
lint = ["ruff >=0.5.5", "pydoclint >=0.5.0"]
docs = [
    "shibuya >=2025.5.30",
    "myst-parser >= 3.0.1",
    "sphinx >= 4.3.2",
    "sphinx-autobuild >=2021.3.14",
    "sphinx-click >=3.0.2",
    "sphinx_mermaid",
    "sphinx_copybutton",
]
mypy = ["mypy >=0.930", "pytest-mock"]
typeguard = ["typeguard >=2.13.3"]
xdoctest = ["xdoctest[colors] >=0.15.10"]


[dependency-groups]
dev = [
    "coverage[toml] >= 6.2",
    "pre-commit >=2.16.0",
    "pre-commit-hooks >=4.6.0",
    "pytest >=6.2.5",
    "pygments >=2.10.0",
    "nox >=2024.04.14",
    "pytest-mock",
]
lint = ["ruff >=0.5.5", "pydoclint >=0.5.0"]
docs = [
    "shibuya >=2025.5.30",
    "myst-parser >= 3.0.1",
    "sphinx >= 4.3.2",
    "sphinx-autobuild >=2021.3.14",
    "sphinx-click >=3.0.2",
    "sphinx_mermaid",
    "sphinx_copybutton",
]
mypy = ["mypy >=0.930", "pytest-mock"]
typeguard = ["typeguard >=2.13.3"]
xdoctest = ["xdoctest[colors] >=0.15.10"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
package = true

[project.scripts]
odoo-data-flow = "odoo_data_flow.__main__:cli"

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["odoo_data_flow", "tests"]
omit = [
    # Exclude the legacy v9 workflow from coverage reports
    "src/odoo_data_flow/lib/workflow/invoice_v9.py",
]

[tool.coverage.report]
show_missing = true
fail_under = 85
exclude_lines = ["pragma: no cover", "if TYPE_CHECKING:"]

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_context = true
# Add this to ignore missing stubs for third-party libraries
ignore_missing_imports = true
check_untyped_defs = true


[[tool.mypy.overrides]]
module = "odoo"
ignore_missing_imports = true


[[tool.mypy.overrides]]
module = "tests.e2e.*"
ignore_errors = true


[tool.ruff]
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "B",   # flake8-bugbear
    "C90", # mccabe
    "D",   # pydocstyle
    "E",   # pycodestyle
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "RUF", # Ruff-specific rules
    "S",   # flake8-bandit
    "UP",  # pyupgrade
    "W",   # pycodestyle
]
ignore = [
    "COM812",
    "COM819",
    "D107",   # undocumented-public-init conflicts with DOC301
    "D206",
    "D300",
    "E111",
    "E114",
    "E117",
    "ISC001",
    "ISC002",
    "Q000",
    "Q001",
    "Q002",
    "Q003",
    "W191",
]
exclude = [
    ".git",
    ".mypy_cache",
    ".nox",
    ".pytest_cache",
    ".venv",
    "__pypackages__",
    "_build",
    "build",
    "dist",
    "docs/conf.py",
    # Add the file you want to ignore to this list
    "pydoclint-baseLine.txt",
    "pydoclint-baseline.txt",
    "uv.lock",
]


[tool.ruff.lint.per-file-ignores]
"*/test_*.py" = ["S101"]
"noxfile.py" = ["S101"]
"**/conftest.py" = ["S101"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pydoclint]
style = 'google'
exclude = '\.git|\.nox|noxfile.py'
arg-type-hints-in-docstring = false
baseline = 'pydoclint-baseline.txt'
auto-generate-baseline = 'True'

[tool.pytest.ini_options]
addopts = "--doctest-modules src tests --ignore=tests/e2e"
doctest_optionflags = "NORMALIZE_WHITESPACE ELLIPSIS IGNORE_EXCEPTION_DETAIL"
